backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
langgraph==0.0.47
langchain-google-genai==0.0.9
langchain==0.1.8
